# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisProgress": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepProgress"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "id",
          "reason",
          "state",
          "error_code",
          "next_attempt_at",
          "steps"
        ],
        "title": "AnalysisProgress",
        "type": "object"
      },
...
      "ApplicationDetail": {
        "additionalProperties": false,
        "properties": {
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/AnalysisProgress"
            },
            "title": "Analyses",
            "type": "array"
          },
          "answers": {
            "items": {
              "discriminator": {
                "mapping": {
                  "date": "#/components/schemas/DateAnswer",
                  "email": "#/components/schemas/EmailAnswer",
                  "file": "#/components/schemas/FileAnswer",
                  "long_text": "#/components/schemas/LongAnswer",
                  "multiple_choice": "#/components/schemas/ChoicesAnswer",
                  "number": "#/components/schemas/NumberAnswer",
                  "short_text": "#/components/schemas/ShortAnswer",
                  "single_choice": "#/components/schemas/ChoiceAnswer",
                  "url": "#/components/schemas/UrlAnswer"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ShortAnswer"
                },
                {
                  "$ref": "#/components/schemas/LongAnswer"
                },
                {
                  "$ref": "#/components/schemas/EmailAnswer"
                },
                {
                  "$ref": "#/components/schemas/NumberAnswer"
                },
                {
                  "$ref": "#/components/schemas/DateAnswer"
                },
                {
                  "$ref": "#/components/schemas/ChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/ChoicesAnswer"
                },
                {
                  "$ref": "#/components/schemas/UrlAnswer"
                },
                {
                  "$ref": "#/components/schemas/FileAnswer"
                }
              ]
            },
            "title": "Answers",
            "type": "array"
          },
          "application": {
            "$ref": "#/components/schemas/ApplicationSummary"
          },
          "effective_evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Evaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "pending_evaluation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Pending Evaluation Ids",
            "type": "array"
          },
          "snapshot": {
            "$ref": "#/components/schemas/PublishedSnapshot"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceVersion"
            },
            "title": "Sources",
            "type": "array"
          },
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/Upload"
            },
            "title": "Uploads",
            "type": "array"
          }
        },
        "required": [
          "snapshot",
          "application",
          "answers"
        ],
        "title": "ApplicationDetail",
        "type": "object"
      },
...
      "ApplicationSummary": {
        "additionalProperties": false,
        "properties": {
          "calculation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Calculation"
              },
              {
                "type": "null"
              }
            ]
          },
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "decision": {
            "enum": [
              "to_review",
              "shortlisted",
              "not_selected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "effective_evaluation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Evaluation Id"
          },
          "eligibility": {
            "anyOf": [
              {
                "enum": [
                  "eligible",
                  "condition_unmet",
                  "needs_review",
                  "not_applicable"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility"
          },
          "evaluation_mode": {
            "anyOf": [
              {
                "enum": [
                  "live",
                  "preloaded"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Mode"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "enum": [
              "real",
              "test"
            ],
            "title": "Mode",
            "type": "string"
          },
          "processing_state": {
            "enum": [
              "queued",
              "collecting",
              "evaluating",
              "completed",
              "completed_partial",
              "failed"
            ],
            "title": "Processing State",
            "type": "string"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "review_revision": {
            "title": "Review Revision",
            "type": "integer"
          },
          "snapshot_id": {
            "format": "uuid",
            "title": "Snapshot Id",
            "type": "string"
          },
          "views": {
            "items": {
              "type": "string"
            },
            "title": "Views",
            "type": "array"
          }
        },
        "required": [
          "id",
          "campaign_id",
          "snapshot_id",
          "mode",
          "contact",
          "received_at",
          "processing_state",
          "decision",
          "review_revision"
        ],
        "title": "ApplicationSummary",
        "type": "object"
      },
...
      "ChoiceAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "single_choice",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "format": "uuid",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "ChoiceAnswer",
        "type": "object"
      },
      "ChoicesAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "multiple_choice",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "title": "Value",
            "type": "array"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "ChoicesAnswer",
        "type": "object"
      },
      "Cleanup": {
        "additionalProperties": false,
        "properties": {
          "application_id": {
            "format": "uuid",
            "title": "Application Id",
            "type": "string"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident": {
            "title": "Incident",
            "type": "boolean"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "state": {
            "enum": [
              "pending",
              "waiting",
              "completed"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "application_id",
          "campaign_id",
          "state",
          "attempts",
          "incident",
          "next_attempt_at",
          "error_code",
          "created_at",
          "completed_at"
        ],
        "title": "Cleanup",
        "type": "object"
      },
...
      "DateAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "date",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "DateAnswer",
        "type": "object"
      },
...
      "EmailAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "email",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "format": "email",
            "maxLength": 320,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "EmailAnswer",
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 100,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "FileAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "file",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5,
            "title": "Value",
            "type": "array"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "FileAnswer",
        "type": "object"
      },
...
      "LongAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "long_text",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "maxLength": 10000,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "LongAnswer",
        "type": "object"
      },
      "NumberAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "number",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "NumberAnswer",
        "type": "object"
      },
...
      "ReviewExpectation": {
        "additionalProperties": false,
        "properties": {
          "effective_evaluation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Evaluation Id"
          },
          "review_revision": {
            "maximum": 9007199254740990.0,
            "minimum": 1.0,
            "title": "Review Revision",
            "type": "integer"
          }
        },
        "required": [
          "review_revision",
          "effective_evaluation_id"
        ],
        "title": "ReviewExpectation",
        "type": "object"
      },
...
      "ShortAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "short_text",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "maxLength": 500,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "ShortAnswer",
        "type": "object"
      },
...
      "UrlAnswer": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "url",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "value": {
            "maxLength": 2048,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "kind",
          "value"
        ],
        "title": "UrlAnswer",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talent Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/applications/{application_id}": {
      "delete": {
        "operationId": "delete_application",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewExpectation"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cleanup"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cleanup"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Delete",
        "tags": [
          "deletions"
        ]
      },
      "get": {
        "operationId": "read_application",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Gone"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Read",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/cleanup-requests": {
      "get": {
        "operationId": "list_cleanup",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cleanup"
                  },
                  "title": "Response List Cleanup",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Listing",
        "tags": [
          "deletions"
        ]
      }
    },
...
    "/api/v1/cleanup-requests/{request_id}": {
      "get": {
        "operationId": "read_cleanup",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cleanup"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Read",
        "tags": [
          "deletions"
        ]
      }
    },
...
  }
}